Fixed race condition in database testing and neatened code
The test was failing because the channel op was occurring before
making the change to known. Since the channel op causes a schedule,
the goroutine was being scheduled without having changed known. We
now change known first.

While I was at it, I did some small refactoring to make db more
self-contained and to remove NewDB, which didn't need to be there
in the first place.

All this needs to be tidied up and documented at some point ¬.¬