Added ErrUpdateNotReady to clarify update failure
It's now much clearer why calling `feed.Update()` had no effect,
if it was prevented because it happened before `feed.Refresh`.

This is done by returning a sentinel error. Note that this error
is also a `net.Error` with obvious behavior, for people who would
rather test behaviour than sentinel errors. For more info on this
thinking,
see:

http://dave.cheney.net/2016/04/27/dont-just-check-errors-handle-them-gracefully

Closes #38