parseTime: text timezones are now parsed correctly
If a format contains the timezone as text,
such as 'Jan 2, 2006 15:04 PM MST',
golang does not automacitally load the timezone
offset. This has to be done manually through
LoadTimezone. The variable
`TimeLayoutsLoadLocation` contains all parse
formats with such a timezone. parseTime now
loads the timezone (if possible) and returns
the time in the correct timezone.

Tests have also been added to test more time
formats. TimeLayouts have also been added
for date that contain both the texual timezone
and offset (such as Mon Jan _2 15:04:05 MST
-0200). In this case, parseTime always uses
the offset, even if it conflicts with the
known offset because it can be ambiguous:
https://news.ycombinator.com/item?id=10199812